Refs #30516 change to uint16_t

+++ Framework/API/inc/MantidAPI/IBoxControllerIO.h
@@ -68,7 +68,7 @@ public:
    * @param blockSize :
    */
   virtual void setDataType(const size_t blockSize, const std::string &typeName,
-                           const uint8_t typeVersion = 2) = 0;
+                           const uint16_t typeVersion = 2) = 0;
   virtual void getDataType(size_t &blockSize, std::string &typeName) const = 0;
 };
 } // namespace API
