+++ Framework/DataObjects/inc/MantidDataObjects/BoxControllerNeXusIO.h
@@ -59,7 +59,7 @@ public:
    * LeanEvent or FatEvent
    * */
   void setDataType(const size_t blockSize, const std::string &typeName,
-                   const uint8_t typeVersion = 2) override;
+                   const uint16_t typeVersion = 2) override;
   void getDataType(size_t &CoordSize, std::string &typeName) const override;
   //------------------------------------------------------------------------------------------------------------------------
   // Auxiliary functions (non-virtual, used for testing)
